[refactor] added aesgcm to primitives
