+++ primitives/aesgcm/ghash.go
@@ -1,4 +1,4 @@
-package aesgcm
+package primitives
 
 // r is a constant used in the multiplication, specific to AES's GF(2^128) field.
 var r = []byte{0xe1, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}
