+++ primitives/ec/symmetric.go
@@ -5,7 +5,7 @@ import (
 	"encoding/base64"
 	"log"
 
-	"github.com/bitcoin-sv/go-sdk/aesgcm"
+	aesgcm "github.com/bitcoin-sv/go-sdk/primitives/aesgcm"
 )
 
 type SymmetricKey struct {
